requirements.txt
huggingface_hub[hf_xet]
numpy
argparse
evaluate
datasets
torch
transformers
scikit-learn
zeus-ml
nvidia-ml-py
accelerate>=0.26.0
wandb
